[rbp] fixed assert condition, stream->time_base.num can be large  enough to overflow the calc as int so do it as int64_t so we do not assert inside ffmpeg.

+++ xbmc/cores/omxplayer/OMXReader.cpp
@@ -590,7 +590,8 @@ void OMXReader::Process()
         if(duration > pStream->duration)
         {
           pStream->duration = duration;
-          duration = m_dllAvUtil.av_rescale_rnd(pStream->duration, pStream->time_base.num * AV_TIME_BASE, pStream->time_base.den, AV_ROUND_NEAR_INF);
+          duration = m_dllAvUtil.av_rescale_rnd(pStream->duration, (int64_t)pStream->time_base.num * AV_TIME_BASE, 
+                                                pStream->time_base.den, AV_ROUND_NEAR_INF);
           if ((m_pFormatContext->duration == (int64_t)AV_NOPTS_VALUE && m_pFormatContext->file_size > 0)
               ||  (m_pFormatContext->duration != (int64_t)AV_NOPTS_VALUE && duration > m_pFormatContext->duration))
             m_pFormatContext->duration = duration;
